+++ examples/__tests__/example1.test.jsx
@@ -1,20 +1,20 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
+import React from "react";
+import ReactDOM from "react-dom";
 import SimpleExample from "../src/examplesFiles/SimpleExample";
-import Example1 from "../src/examplesFiles/Example1";
-import Example2 from "../src/examplesFiles/Example2";
+import FewArrows from "../src/examplesFiles/FewArrows";
+import CustomizeArrow from "../src/examplesFiles/CustomizeArrow";
 
-test("SimpleExample renders without crashing",()=>{
-  const div = document.createElement('div')
-  ReactDOM.render(<SimpleExample/>,div)
-})
+test("SimpleExample renders without crashing", () => {
+  const div = document.createElement("div");
+  ReactDOM.render(<SimpleExample />, div);
+});
 
-test("Example1 renders without crashing",()=>{
-  const div = document.createElement('div')
-  ReactDOM.render(<Example1/>,div)
-})
+test("FewArrows renders without crashing", () => {
+  const div = document.createElement("div");
+  ReactDOM.render(<FewArrows />, div);
+});
 //
-// test("Example2 renders without crashing",()=>{
+// test("CustomizeArrow renders without crashing",()=>{
 //   const div = document.createElement('div')
-//   ReactDOM.render(<Example2/>,div)
+//   ReactDOM.render(<CustomizeArrow/>,div)
 // })