+++ examples/src/examplesFiles/FewArrows.jsx
@@ -29,7 +29,7 @@ const boxStyle = {
   height: "30px",
 };
 
-const Example1 = () => {
+const FewArrows = () => {
   const [, setRender] = useState({});
   const forceRerender = () => setRender({});
 
@@ -107,4 +107,4 @@ const Example1 = () => {
   );
 };
 
-export default Example1;
+export default FewArrows;